refactor: extract LogBuffer to utils/log-buffer.ts
Move LogBuffer class out of runtime.ts into its own utility file
to reduce runtime.ts bloat.
